requirements.txt
python-dateutil
python-memcached==1.57
chardet
six
Pillow==4.3.0
Django==1.8.18
django-compressor==1.4
django-post-office==2.0.3
django-statici18n==1.1.2
djangorestframework==3.3.1
git+git://github.com/haiwen/django-constance.git@8508ff29141732190faff51d5c2b5474da297732#egg=django-constance[database]
openpyxl==2.3.0
pytz==2015.7
django-formtools
qrcode
requests
requests_oauthlib==0.8.0
django-webpack-loader==0.6.0
